Update weapon_system_api.lua
Hi, the current weapon system API links the AIM-9L to the AIM-9M's CSLID, preventing the 9Ls from being shown correctly in the MFD when loaded. This update corrects the issue plus I threw in a number of other models of Aim-9s in case they are added to the a-29b lua later.