Engine management improvements
* Engine:

Engine starting/shutting down will now take into consideration the switches positions, whether the starter is on, the ignition status, the power supply and the engine rpm.

It's now possible to execute dry startups, startups with PMU in manual and automatic reignition.

The engine can now start with main pumps in off (mechanical pump mode).

Using the starter now heavily drains the battery if no external power is supplied. This should be enough for one startup in flight, or one in the ground.

Added the ENG MAN alarm.